server/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9  # For PostgreSQL
openai==1.2.0           # For OpenRouter API integration
pytest==7.4.3           # For testing